REFACTOR : 상품 상세 정보와 입찰 api 합체
* 한 객체에서 처리합니다.
