+++ app/src/main/kotlin/com/fakedevelopers/bidderbidder/api/di/ApiModule.kt
@@ -1,13 +1,11 @@
 package com.fakedevelopers.bidderbidder.api.di
 
 import com.fakedevelopers.bidderbidder.api.data.Constants.Companion.BASE_URL
-import com.fakedevelopers.bidderbidder.api.repository.ProductBidRepository
 import com.fakedevelopers.bidderbidder.api.repository.ProductDetailRepository
 import com.fakedevelopers.bidderbidder.api.repository.ProductListRepository
 import com.fakedevelopers.bidderbidder.api.repository.ProductRegistrationRepository
 import com.fakedevelopers.bidderbidder.api.repository.SigninGoogleRepository
 import com.fakedevelopers.bidderbidder.api.repository.UserLoginRepository
-import com.fakedevelopers.bidderbidder.api.service.ProductBidService
 import com.fakedevelopers.bidderbidder.api.service.ProductDetailService
 import com.fakedevelopers.bidderbidder.api.service.ProductListService
 import com.fakedevelopers.bidderbidder.api.service.ProductRegistrationService
@@ -114,7 +112,7 @@ object ApiModule {
     fun provideSigninGoogleRepository(service: SigninGoogleService): SigninGoogleRepository =
         SigninGoogleRepository(service)
 
-    // 상품 상세 정보
+    // 상품 상세 정보, 입찰
     @Singleton
     @Provides
     fun provideProductDetailService(retrofit: Retrofit): ProductDetailService =
@@ -124,15 +122,4 @@ object ApiModule {
     @Provides
     fun provideProductDetailRepository(service: ProductDetailService): ProductDetailRepository =
         ProductDetailRepository(service)
-
-    // 상품 입찰
-    @Singleton
-    @Provides
-    fun provideProductBidService(retrofit: Retrofit): ProductBidService =
-        retrofit.create(ProductBidService::class.java)
-
-    @Singleton
-    @Provides
-    fun provideProductBidRepository(service: ProductBidService): ProductBidRepository =
-        ProductBidRepository(service)
 }
