+++ app/src/main/kotlin/com/fakedevelopers/bidderbidder/api/service/ProductDetailService.kt
@@ -2,7 +2,10 @@ package com.fakedevelopers.bidderbidder.api.service
 
 import com.fakedevelopers.bidderbidder.ui.product_detail.ProductDetailDto
 import retrofit2.Response
+import retrofit2.http.Field
+import retrofit2.http.FormUrlEncoded
 import retrofit2.http.GET
+import retrofit2.http.POST
 import retrofit2.http.Path
 
 interface ProductDetailService {
@@ -10,4 +13,12 @@ interface ProductDetailService {
     suspend fun getProductDetail(
         @Path("productId") productId: Long
     ): Response<ProductDetailDto>
+
+    @FormUrlEncoded
+    @POST("product/{productId}/bid")
+    suspend fun postProductBid(
+        @Path("productId") productId: Long,
+        @Field("userId") userId: Long,
+        @Field("bid") bid: Long
+    ): Response<String>
 }
