+++ app/src/main/kotlin/com/fakedevelopers/bidderbidder/ui/product_detail/ProductDetailViewModel.kt
@@ -3,7 +3,6 @@ package com.fakedevelopers.bidderbidder.ui.product_detail
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.fakedevelopers.bidderbidder.api.data.Constants.Companion.dec
-import com.fakedevelopers.bidderbidder.api.repository.ProductBidRepository
 import com.fakedevelopers.bidderbidder.api.repository.ProductDetailRepository
 import com.fakedevelopers.bidderbidder.ui.util.ExpirationTimerTask
 import com.fakedevelopers.bidderbidder.ui.util.MutableEventFlow
@@ -18,8 +17,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class ProductDetailViewModel @Inject constructor(
-    private val productDetailRepository: ProductDetailRepository,
-    private val productBidRepository: ProductBidRepository
+    private val repository: ProductDetailRepository
 ) : ViewModel() {
 
     private val _productDetailResponse = MutableEventFlow<Response<ProductDetailDto>>()
@@ -87,7 +85,7 @@ class ProductDetailViewModel @Inject constructor(
     fun productDetailRequest(productId: Long) {
         this.productId = productId
         viewModelScope.launch {
-            _productDetailResponse.emit(productDetailRepository.getProductDetail(productId))
+            _productDetailResponse.emit(repository.getProductDetail(productId))
         }
     }
 
@@ -254,7 +252,7 @@ class ProductDetailViewModel @Inject constructor(
         // 입찰가 조작을 막고 api 요청
         setBiddingEnabled(false)
         viewModelScope.launch {
-            _productBidResponse.emit(productBidRepository.postProductBid(productId, requestUserId, requestBid))
+            _productBidResponse.emit(repository.postProductBid(productId, requestUserId, requestBid))
         }
     }
